Created update 1.4.0
Split off monolithic menu library functions into individual files, reducing RAM consumption significantly (2022-12-15).
Increased aggressive GC threshold from 4K to 16K due to RAM usage savings from modularization (2022-12-15).
Changed charge error messages to reflect if recovery is enabled (i.e. "paused" vs. "stopped") (2022-12-18).
Fixed issue where precharge was not subject to the safety time limit (2022-12-18).
Added cumulative charge/energy display for the current charge session (2022-12-24).
Updated free memory counter to specify count in bytes (2022-12-24).
Changed how the UI calculates when to show different statusbar messages (2012-12-24).
