+++ source/lua/user/DC4S/lib/DC4S-cfgExtTemp.lua
@@ -1,29 +1,8 @@
---[[DingoCharge for Shizuku Platform - Menu Library 2
-https://ripitapart.com December 13, 2022.
+--[[DingoCharge for Shizuku Platform - External Temperature Sensor Configuration Menu
+https://ripitapart.com December 15, 2022.
 
 Version history:
-1.0.0: Initial public release (2022-06-30).
-1.1.0: Fixed issue where setting cell count does not update precharge voltage. (2022-07-22).
-       Fixed issue where precharge voltage did not display (correctly) during charge. (2022-10-12).
-       Added CC fallback when in CV mode and charge current overshoots too much (2022-10-12).
-       Added an option to display system temperature in Fahrenheit (2022-10-12).
-       Added 2.5V/cell and 8S cell configurations (2022-10-13).
-1.1.1: Fixed issue where some chargers' current-limiting conflicted with CV control loop (2022-10-15).
-1.1.2: Fixed issue where setting 8S configuration would result in a Config Error message (2022-10-20).
-1.2.0: Added prompt to retry the compatibility test if Vbus voltage is not present, instead of outright failing (2022-11-06).
-       Added CC deadband threshold tweaks to fix an issue where setting charge current overwrites the user's defaults (2022-11-06).
-       Fixed issue where double-tapping Select key in "Advanced... > Chg Reg Deadband" menu does not go up a level (2022-11-06).
-1.3.0: Fixed issue where 3.65Vpc was considered standard Li-ion in terms of precharge voltage instead of LiFePO4 (2022-11-17).
-       Added test to verify configuration immediately upon startup (2022-12-12).
-       Added memory cleanup routine after reading PDOs from adapter (2022-12-12).
-       Changed internal version format (2022-12-12).
-       Added statusbar override support for charge termination/faults (2022-12-13).
-       Added support for TMP3x/LM35/LM50 external temperature sensor on D+ pin (2022-12-13).
-       Added optional over/undertemperature protection when using external sensor (2022-12-13).
-       Added charge timeout protection (2022-12-13).
-       Fixed issue where resuming session timer counts time while timer was stopped (2022-12-13).
-       Added second menu library file due to RAM space exhaustion (2022-12-13).
-       Decreased aggressive GC threshold from 16K to 4K but added more forced GCs to mitigate RAM exhaustion (2022-12-13).]]
+1.4.0: Split off monolithic menu library functions into individual files (2022-12-15).]]
 
 function cfgExtTemp()
   local cfgExtTempMenuEntryExtEnable = " "
@@ -193,6 +172,7 @@ function cfgExtTemp()
         timer = timer - 1
          screen.popHint(string.format("Ext Temp: %+.1fC", readExternalTemperatureCelsius()), 100)
       end
+      timer = nil
     elseif cfgExtTempSel == 10 then
       if (screen.popYesOrNo("Restore defaults?", color.yellow)) then
         setExternalTemperatureDefaults()
@@ -202,45 +182,14 @@ function cfgExtTemp()
       break
     end
   end
-  collectgarbage("collect") -- clean up memory
-end
-
-function cfgTimeLimit()
-  local cfgTimeLimitSel = 0
-  local cfgTimeLimitText = " "
-  local tmpTimeLimit = 0
-  while true do
-    screen.clear()
-    if timeLimitHours == 0 then
-      cfgTimeLimitText = "Keep Current (Disabled)"
-    else
-      cfgTimeLimitText = string.format("Keep Current (%dh)", timeLimitHours)
-    end
-    cfgTimeLimitSel = screen.popMenu{cfgTimeLimitText, "Disable Time Limit", "Set Time Limit...", "Restore Defaults"}
-    if cfgTimeLimitSel == 1 then
-      timeLimitHours = 0
-      break
-    elseif cfgTimeLimitSel == 2 then
-      -- Tens
-      tmpTimeLimit = 10 * screen.popMenu{"0xh","1xh","2xh","3xh","4xh","5xh","6xh","7xh","8xh","9xh"}
-      -- Ones
-      tmpTimeLimit = tmpTimeLimit + (screen.popMenu({string.format("%dh",tmpTimeLimit),string.format("%dh",tmpTimeLimit + 1),string.format("%dh",tmpTimeLimit + 2),string.format("%dh",tmpTimeLimit + 3),string.format("%dh",tmpTimeLimit + 4),string.format("%dh",tmpTimeLimit + 5),string.format("%dh",tmpTimeLimit + 6),string.format("%dh",tmpTimeLimit + 7),string.format("%dh",tmpTimeLimit + 8),string.format("%dh",tmpTimeLimit + 9)}))
-      timeLimitHours = tmpTimeLimit
-      collectgarbage("collect")
-      break
-    elseif cfgTimeLimitSel == 3 then
-      if (screen.popYesOrNo("Restore defaults?", color.yellow)) then
-        setTimeLimitDefaults()
-        screen.popHint("Defaults Restored", 1000)
-      end
-    else
-      break
-    end
-  end
-  if timeLimitHours == 0 then
-    screen.popHint("Disabled", 1000)
-  else
-    screen.popHint(string.format("%dh", timeLimitHours), 1000)
-  end
+  -- discard temporary variables
+  cfgExtTempMenuEntryExtEnable = nil
+  cfgExtTempMenuEntryOtp = nil
+  cfgExtTempMenuEntryOtpRecovery = nil
+  cfgExtTempMenuEntryUtp = nil
+  cfgExtTempMenuEntryUtpRecovery = nil
+  cfgExtTempSel = nil
+  cfgExtTempSelSubmenu = nil
+  cfgExtTempSelSubmenuText = nil
   collectgarbage("collect") -- clean up memory
 end
